invoice-processor/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
pandas==2.1.3
python-dateutil==2.8.2
pydantic==2.5.2
PyMuPDF==1.23.7 